pyproject.toml
[project]
name = "learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.4",
    "notebook>=7.2.2",
    "tortoise-orm>=0.21.7",
]
